:adhesive_bandage: fixed minor sintax error

+++ src/db/db_empresa.sql
@@ -6,5 +6,5 @@ CREATE TABLE TB_Colaborador(
     colab_nome VARCHAR(20) NOT NULL,
     colab_sobrenome VARCHAR(40) NOT NULL,
     colab_cpf VARCHAR(11) NOT NULL,
-    colab_salLiq FLOAT NOT NULL
+    colab_salLiq DECIMAL(7,2) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 DEFAULT COLLATE utf8_unicode_ci;
